strategies do not store cache on self
Since strategies are now instance of class, and not constants,
thread-safety must be accounted for.

The usage of `self.cache` is, in this regard, an issue. To resolve it,
I reuse the same behavior as before in `taal.Translator`: cache is
generated once, and passed around as argument in every subsequent call.